# openapi.yaml
openapi: 3.0.3
info:
  title: YieldScan Backend API
  description: |
    API for managing crypto assets, their yield-generating steps, and retrieving APY data.
    This API supports protocol-based step organization with comprehensive asset and step management capabilities.
  version: 1.0.0
  contact:
    name: YieldScan Team
servers:
  - url: http://localhost:4023/api
    description: Local development server
  - url: https://api.yieldscan.com/api
    description: Production server
paths:
...
  /chains/{id}:
    get:
      tags:
        - Chains
      summary: Get chain by ID
      description: Retrieve a specific chain by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Chain ID
          schema:
            type: string
            example: 684a8cf6db44b98888f22f5f
        - name: includeDisabled
          in: query
          description: Include disabled chains in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  chain:
                    $ref: '#/components/schemas/Chain'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Chains
      summary: Update chain
      description: Update an existing chain
      parameters:
        - name: id
          in: path
          required: true
          description: Chain ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChainRequest'
      responses:
        '200':
          description: Chain updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Chain updated successfully
                  chain:
                    $ref: '#/components/schemas/Chain'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Chains
      summary: Delete chain
      description: Delete a chain by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Chain ID
          schema:
            type: string
      responses:
        '200':
          description: Chain deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Chain deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /protocols/{id}:
    get:
      tags:
        - Protocols
      summary: Get protocol by ID
      description: Retrieve a specific protocol by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Protocol ID
          schema:
            type: string
            example: 684a8cf6db44b98888f22f5f
        - name: includeDisabled
          in: query
          description: Include disabled protocols in results
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  protocol:
                    $ref: '#/components/schemas/Protocol'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Protocols
      summary: Update protocol
      description: Update an existing protocol
      parameters:
        - name: id
          in: path
          required: true
          description: Protocol ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProtocolRequest'
      responses:
        '200':
          description: Protocol updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Protocol updated successfully
                  protocol:
                    $ref: '#/components/schemas/Protocol'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Protocols
      summary: Delete protocol
      description: Delete a protocol by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Protocol ID
          schema:
            type: string
      responses:
        '200':
          description: Protocol deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Protocol deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Chain:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        chainId:
          type: integer
          example: 1
        name:
          type: string
          example: Ethereum
        shortName:
          type: string
          example: ETH
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/info/logo.png"
        rpcUrl:
          type: string
          example: "https://eth.llamarpc.com"
        explorerUrl:
          type: string
          example: "https://etherscan.io"
        symbol:
          type: string
          example: ETH
        decimals:
          type: integer
          example: 18
        enabled:
          type: boolean
          example: true
        testnet:
          type: boolean
          nullable: true
          example: false
        description:
          type: string
          nullable: true
          example: "Ethereum mainnet"
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
...
    UpdateChainRequest:
      type: object
      properties:
        chainId:
          type: integer
        name:
          type: string
        shortName:
          type: string
        image:
          type: string
        rpcUrl:
          type: string
        explorerUrl:
          type: string
        symbol:
          type: string
        decimals:
          type: integer
        enabled:
          type: boolean
        testnet:
          type: boolean
          nullable: true
        description:
          type: string
          nullable: true
    Protocol:
      type: object
      properties:
        _id:
          type: string
          example: 684a8cf6db44b98888f22f5f
        name:
          type: string
          example: Aave
        displayName:
          type: string
          example: Aave
        image:
          type: string
          example: "https://cdn.jsdelivr.net/gh/trustwallet/assets/blockchains/ethereum/assets/0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9/logo.png"
        website:
          type: string
          nullable: true
          example: "https://aave.com"
        description:
          type: string
          nullable: true
          example: "Open source and non-custodial liquidity protocol for earning interest on deposits and borrowing assets"
        category:
          type: string
          nullable: true
          example: "Lending"
        enabled:
          type: boolean
          example: true
        supportedChains:
          type: array
          items:
            type: integer
          nullable: true
          example: [1, 137, 42161, 8453, 56]
        createdAt:
          type: string
          format: date-time
          example: "2025-06-16T08:11:29.443Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-06-16T08:25:35.229Z"
...
    UpdateProtocolRequest:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        image:
          type: string
        website:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        enabled:
          type: boolean
        supportedChains:
          type: array
          items:
            type: integer
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        details:
          type: string
          example: "Missing required field: token"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Missing required fields"
            details: "assetId, order, title, description, chain, functionConfig are required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Asset not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Failed to fetch assets"
            details: "Database connection error"
...